Created styles directory and moved all styled files to this directory
