+++ src/styles/StyledApp.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import App from './App';
+import App from '../App';
 
 const StyledApp = styled(App)`
   display: flex;
